Web/final_project/requirements.txt
PyYAML~=6.0.1
selenium~=4.11.2
pytest~=7.4.0
requests~=2.31.0